WIP clearing flags out
I've decided to abandon this line of work.

A particular block might be false, but a remote change somewhere else in the
touch might then change it to be true. The block will continue to think it's
false until traversed again by a visitor. Until then the information is out-of-
date and misleading.

Initial attempts to clear flags on changes were abandoned. Clearing flags on
all blocks in response to a simple change (e.g. a call changing elsewhere) will
be very costly.

This has made me think that this is the wrong strategy for what I'm trying to
accomplish. Truth is a global property of an entire touch, not a local property
of an individual block.