Performance optimisation for music analyses
Music matchers used to take a row and return whether it matched. This is
the most logical interface, but given that the row needs to be converted
to a string for matching it means we need to call stringFromRow() for
every individual pattern being matched.

This commit modifies matchers to take a row string instead for matching.
The call to stringFromRow() is moved into the Music visitor.