Removed Stedman-specific propagation code from AbstractContainer
I'd designed `AbstractContainer` to make it easy for derived classes to
customise how the first block is propagated. This was needed for Stedman
(to support starts), but adds extra complexity to the `AbstractContainer`
itself.

This commit simplifies `AbstractContainer` by removing this extensibility.
Instead we override the entirety of `propagateBlocks` in the derived
`Touch`. We need to be careful doing this, but it's thoroughly covered by
existing tests and the simplification is beneficial when supporting other
prickers / methods.