Use environment variable for S3 bucket name
### Why?

Deployment to an S3 bucket is handled by AWS CodeBuild with credentials
managed by an AWS role. This infrastructure used to be managed by raw
CloudFormation but I migrated it to be managed using the AWS CDK. The
bucket name is now autogenerated with a random slug.

The bucket name is already available: we just need to use it.