Don't bother to generate courses when start and end row are the same
### Why?

We end up with one course, but it doesn't show up in the list of courses
because it has no calls. Despite this it's still possible to click on it
and open it, leading to errors. The window also says "1 courses" which
is pretty misleading.

Let's just clear the list of courses for this state.