Refactor: use class properties to update UI
### Why?

We currently call `render()` from a lot of handlers. This redraws all
elements on the page regardless of whether or not they've changed. As we
add additional functionality we don't necessarily want this.

Instead let's define `get` / `set` accessors for properties that match
elements in the UI. When we set a property it should update the UI
automatically.

So far it seems like this is beneficial: it makes it easier to ensure
that downstream effects (updating the UI and/or updating other values)
are triggered when necessary. I'm slightly nervous it'll be trickier to
avoid triggering these effects multiple times by mistake.

### What?

* add accessors for many private class properties
* migrate `redraw()` content into `set` accessors so the UI is kept in
  sync with any changes automatically