Rewrite `createTable()` algorithm
### Why?

I attempted to optimise table creation by storing a list of the rows to
process at the next iteration (rather than searching the entire table
each time to find them again). Performance of the new algorithm was
roughly the same as the old, but I think it's slightly clearer what's
going on so have decided to keep it.

### What?

* rework `createTable()` to keep track of the rows to process in the
  next iteration
* add timing output to `createDataTables.js` script

### Performance comparison

Old algorithm: 1m15.738s
New algorithm: 1m13.319s