Import TypeScript types using `type` keyword
### Why?

From the [`typescript-eslint` documentation]:

> TypeScript allows specifying a type keyword on imports to indicate
> that the export exists only in the type system, not at runtime. This
> allows transpilers to drop imports without knowing the types of the
> dependencies.

[`typescript-eslint` documentation]: https://typescript-eslint.io/rules/consistent-type-imports/

This eliminates a big stack of ESLint warnings about dependency cycles.
Cycles involving code are a big problem: if modules aren't packed in the
bundle in the right way then something might break. Cycles involving
types are not really an issue. The compiler seems to be able to handle
them and they don't affect the compiled bundle.

### What?

* enable `@typescript-eslint/consistent-type-imports` rule
* disable `import/consistent-type-specifier-style` because it's useful
  to be able to write imports in the most concise way
* make many many fixes