Remove `fromString` methods from courses and touches
### Why?

As reported by [`dpdm`](https://github.com/acrazing/dpdm):

```
• Circular Dependencies
  1) src/grandsire/Course/index.ts -> src/grandsire/Course/Course.ts -> src/grandsire/Parser.ts
  2) src/grandsire/Parser.ts -> src/grandsire/Touch/index.ts -> src/grandsire/Touch/Touch.ts
  3) src/stedman/Course/index.ts -> src/stedman/Course/Course.ts -> src/stedman/Parser.ts
  4) src/stedman/Parser.ts -> src/stedman/Touch/index.ts -> src/stedman/Touch/Touch.ts
```

A `Parser` has to be able to create `Course` and `Touch` objects: that's
its job. If `Course.fromString()` uses a `Parser` then we create a
circular dependency.

The `.fromString()` methods are _nice_ but they're not at all essential.
Using the `Parser` directly results in this diff for the grandsire
pricker:

```diff
-            newTouch = Touch.fromString(input);
+            newTouch = new Parser().parseTouch(input);
```

### What?

* remove `.fromString()` from `Course` and `Touch` classes for Grandsire
  and Stedman prickers
* replace all usage with direct calls to the relevant `Parser`