Reorganise searching code... again
Follows on from 5d39a06

### Why?

The search steps and their callings are indeed specific to methods.
Unfortunately that's a bad place to generate search transpositions
because the easiest way to do that is to create a `Course` and configure
it with each calling. We can't do that from inside methods because
courses depend on methods already, i.e. we'd create a circular
dependency.

We can work round this by only storing the calling string on each
method. The actual transposition generation moves back to a
`createTranspositions()` function... but that's now generic to any
method (even Grandsire).

### What?

* add `searchCallingStrings` to `AbstractMethod` and `Erin`
  (which works differently)
* reintroduce `createTranspositions()` which takes a course and an array
  of calling strings in order to produce the transpositions
* tweak `Calling` so it doesn't create courses but updates them instead
  (removing one location where `Stedman` is hardcoded)
* update `createTable()` and `search()` accordingly