Modify `Uint4Table` to store data in unpacked form
### Why?

Our `Uint4Table` packs data for two rows into each byte of the table.
This is great for download: the table for Stedman Cinques drops from
nearly 40MB to only 20MB.

We currently read from and write to the packed table, performing bitwise
operations to update it directly. This is tricky to maintain, and it
will be awkward to build and test a `Uint6Table` for other methods.
Let's instead perform all the packing and unpacking in one go and store
the unpacked table inside the class.

### What?

* unpack data on construction and repack it when accessing `.data`
* update `getValue()` and `setValue()` to be simple array accesses

### Performance

It's not obvious that this will perform better or worse than the old
implementation: we're avoiding lots of bitwise operations but are using
more memory (and accessing data across that memory).

I usually measure a ≈20% increase in the time taken to build data tables
(`createTable` using the new approach). I _have_ measured a similar
increase when reading table data, but rerunning the experiment today I
find no change in speed.

In any case a 20% speed reduction is pretty insignificant. I am able to
generate 155,077 touches in <100ms on this machine - far more touches
than could ever be reviewed manually. It's worth paying the price for an
increase in readability.