Create separate suite of packing algorithms
### Why?

Our current plan is to maintain different table classes for different
types of packing - we have `Uint4Table` and `Uint6Table` so far.
This has one problem: when we're generating tables we don't yet know the
size of the values we want to store (and hence which table to use).
The choice of table wants to leak out rather than stay hidden.

It would be better to encapsulate the choice of packing algorithm within
a more-generic `Table` class. A `Table` can choose the algorithm itself
depending on the values it stores or the size of the data it's given.

### What?

* create a family of packing algorithms:
  * `AbstractPacking` base class
  * `Uint4Packing`, `Uint6Packing` and `Uint8Packing` (new!)
* create a `Table` class that provides the same interface as existing
  tables but chooses the packing algorithm dynamically