Rework packing algorithms to store a value one higher
### Why?

In 26810a0 we noted:

> The table for Stedman Cinques contains 250 rows that are 16 steps from
> rounds, e.g. `980E6342571`.

The 4 bits we've allocated with our `Uint4Packing` only provide space to
store values from 0 to 15, which means we need to adopt `Uint6Packing`
and add 10MB to the size of the Stedman Cinques data table.

We can store one additional value in each table by noting that there is
only one row that is zero steps from rounds: rounds itself. If we ignore
the first entry in the array then we're actually trying to store values
from 1 to 16. These fit in the range 0 to 15 if we subtract one.

### What?

* update packing algorithms to subtract one when packing values and add
  it back when unpacking
* take the opportunity to switch the rows in `Uint4Packing` so they're
  laid out sequentially in memory (at least in theory)
* rework `Uint4Packing` algorithm so it looks more like `Uint6Packing`
* regenerate data tables

### How was this tested?

We expect the following changes:

1. nibbles should be swapped over in each byte of a data table
2. values should be decreased by one (except the first entry, still 0)
3. (for Cinques only) 250 rows should be increased by one

I wrote a simple Python script to perform steps 1 and 2 above and used
it to verify that newly-generated tables were transformed as expected.

```python
FILE = 'data/stedman.7.dat'

with open(FILE, mode='rb', buffering=0) as f:
    data = bytearray(f.read())

for i, byte in enumerate(data):
    first = ((byte >> 4) & 0xf) - 1
    second = (byte & 0xf) - 1

    if i == 0:
        second = 0

    data[i] = (second << 4) | first

with open(FILE, mode='wb', buffering=0) as f:
    f.write(data)
```

This doesn't work for Cinques, so I've manually verified some searches.